+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ### 1.1 (Unreleased)
 * Connected Recognition Awards are handled in a special way.
-
+* No longer verify SSL when checking Github for updated version.
 
 ### 1.0 (May 3rd, 2013)